+++ NEWS.md
@@ -1,3 +1,12 @@
+# pixeltrix 0.1.1
+
+* Added `draw_pixels()` to plot the matrix with `image()` to the plotting window (#12).
+* Added `frame_pixels()` and `gif_pixels()` to capture successive 'frames' of an animation and write it to a gif (#13).
+* Suggested {gifski} for creating gifs (towards #15).
+* Provided to the user some instruction messages when they enter interactive mode.
+* Improved code commentary.
+* Added two matrix outputs as example datasets (Pokemon and Mario).
+
 # pixeltrix 0.1.0
 
 * Added basic error tests.